+++ packages/core/solidity/src/erc1155.ts
@@ -10,6 +10,10 @@ import { withCommonDefaults, defaults as commonDefaults } from './common-options
 import { setUpgradeable } from './set-upgradeable';
 import { setInfo } from './set-info';
 import { printContract } from './print';
+import { addCrosschainLinked } from './set-crosschain-linked';
+
+export const crossChainBridgingOptions = [false, 'erc7786native'] as const;
+export type CrossChainBridging = (typeof crossChainBridgingOptions)[number];
 
 export interface ERC1155Options extends CommonOptions {
   name: string;
@@ -19,6 +23,8 @@ export interface ERC1155Options extends CommonOptions {
   mintable?: boolean;
   supply?: boolean;
   updatableUri?: boolean;
+  crossChainBridging?: CrossChainBridging;
+  crossChainLinkAllowOverride?: boolean;
 }
 
 export const defaults: Required<ERC1155Options> = {
@@ -30,6 +36,8 @@ export const defaults: Required<ERC1155Options> = {
   mintable: false,
   supply: false,
   updatableUri: true,
+  crossChainBridging: false,
+  crossChainLinkAllowOverride: false,
 } as const;
 
 function withDefaults(opts: ERC1155Options): Required<ERC1155Options> {
@@ -41,6 +49,8 @@ function withDefaults(opts: ERC1155Options): Required<ERC1155Options> {
     mintable: opts.mintable ?? defaults.mintable,
     supply: opts.supply ?? defaults.supply,
     updatableUri: opts.updatableUri ?? defaults.updatableUri,
+    crossChainBridging: opts.crossChainBridging ?? defaults.crossChainBridging,
+    crossChainLinkAllowOverride: opts.crossChainLinkAllowOverride ?? defaults.crossChainLinkAllowOverride,
   };
 }
 
@@ -49,7 +59,13 @@ export function printERC1155(opts: ERC1155Options = defaults): string {
 }
 
 export function isAccessControlRequired(opts: Partial<ERC1155Options>): boolean {
-  return opts.mintable || opts.pausable || opts.updatableUri !== false || opts.upgradeable === 'uups';
+  return (
+    opts.mintable ||
+    opts.pausable ||
+    opts.updatableUri !== false ||
+    opts.upgradeable === 'uups' ||
+    opts.crossChainBridging === 'erc7786native'
+  );
 }
 
 export function buildERC1155(opts: ERC1155Options): Contract {
@@ -60,6 +76,10 @@ export function buildERC1155(opts: ERC1155Options): Contract {
   const { access, upgradeable, info } = allOpts;
 
   addBase(c, allOpts.uri);
+
+  if (allOpts.crossChainBridging) {
+    addCrossChainBridging(c, allOpts.crossChainBridging, allOpts.crossChainLinkAllowOverride, access);
+  }
 
   if (allOpts.updatableUri) {
     addSetUri(c, access);
@@ -122,6 +142,35 @@ function addMintable(c: ContractBuilder, access: Access) {
   requireAccessControl(c, functions.mintBatch, access, 'MINTER', 'minter');
   c.addFunctionCode('_mint(account, id, amount, data);', functions.mint);
   c.addFunctionCode('_mintBatch(to, ids, amounts, data);', functions.mintBatch);
+}
+
+function addCrossChainBridging(
+  c: ContractBuilder,
+  crossChainBridging: 'erc7786native',
+  crossChainLinkAllowOverride: boolean,
+  access: Access,
+) {
+  switch (crossChainBridging) {
+    case 'erc7786native':
+      addERC1155Crosschain(c, crossChainLinkAllowOverride, access);
+      break;
+    default: {
+      const _: never = crossChainBridging;
+      throw new Error('Unknown value for `crossChainBridging`');
+    }
+  }
+}
+
+function addERC1155Crosschain(c: ContractBuilder, crossChainLinkAllowOverride: boolean, access: Access) {
+  addCrosschainLinked(
+    c,
+    {
+      name: 'ERC1155Crosschain',
+      path: '@openzeppelin/contracts/token/ERC1155/extensions/ERC1155Crosschain.sol',
+    },
+    crossChainLinkAllowOverride,
+    access,
+  );
 }
 
 function addSetUri(c: ContractBuilder, access: Access) {